% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_an.R
\name{ac}
\alias{ac}
\title{ac - short-cut to 'as.character'}
\usage{
ac(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
character type vector
}
\description{
Short-cut wrapper for (\code{\link[base]{as.numeric}}
}
\examples{

x <- 1:4
ac(x)

}
