% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addAlpha.R
\name{addAlpha}
\alias{addAlpha}
\title{Add alpha channel (transparency) to colors}
\usage{
addAlpha(COLORS, ALPHA)
}
\arguments{
\item{COLORS}{Vector of any of the three kinds of R color specifications,
i.e., either a color name (as listed by colors()),
a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb),
or a positive integer i meaning palette()[i].}

\item{ALPHA}{A value (between 0 and 1) indicating the alpha channel
(opacity) value.}
}
\description{
Takes a vector of colors and adds an alpha channel
at the given level of transparency.
}
\examples{
# Make background image
x <- seq(-180, 180,, 30)
y <- seq(-90, 90,, 30)
grd <- expand.grid(x=x,y=y)
z <- sqrt(grd$x^2+grd$y^2)
dim(z) <- c(length(x), length(y))
pal <- colorRampPalette(c(rgb(1,1,1), rgb(0,0,0)))
COLORS <- pal(20)
image(x,y,z, col=COLORS)

# Add semi-transparent layer
z2 <- grd$x^2+grd$y
dim(z2) <- c(length(x), length(y))
pal <- colorRampPalette(c(rgb(0.5,1,0), rgb(0,1,1), rgb(1,1,1)))
COLORS <- addAlpha(pal(20), 0.4) # alpha chanel equals 0.4
image(x,y,z2, col=COLORS, add=TRUE)
}
\keyword{color}

