% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_an.R
\name{an}
\alias{an}
\title{an - short-cut to 'as.numeric'}
\usage{
an(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
numeric type vector
}
\description{
Short-cut wrapper for (\code{\link[base]{as.character}}
}
\examples{

x <- c("1", "2", "3", "4")
an(x)

}
