% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{bioEnv}
\alias{bioEnv}
\title{Clarke and Ainsworth's BIO-ENV routine}
\usage{
bioEnv(fix.mat, var.mat, fix.dist.method = "bray",
  var.dist.method = "euclidean", scale.fix = FALSE, scale.var = TRUE,
  output.best = 10, var.max = ncol(var.mat))
}
\arguments{
\item{fix.mat}{The "fixed" matrix of community or environmental sample by variable values}

\item{var.mat}{A "variable" matrix of community or environmental sample by variable values}

\item{fix.dist.method}{The method of calculating dissimilarity indices bewteen samples in the fixed
matrix (Uses the \code{\link[vegan]{vegdist}} function from the vegan package to calculate distance matrices. See
the documentation for available methods.). Defaults to Bray-Curtis dissimularity \code{"bray"}.}

\item{var.dist.method}{The method of calculating dissimilarity indices bewteen samples in the variable
matrix. Defaults to Euclidean dissimularity \code{"euclidean"}.}

\item{scale.fix}{Logical. Should fixed matrix be centered and scaled (Defaults to \code{FALSE},
recommended for biologic data).}

\item{scale.var}{Logical. Should fixed matrix be centered and scaled (Defaults to \code{TRUE},
recommended for environmental data to correct for differing units between variables).}

\item{output.best}{Number of best combinations to return in the results object (Default=10).}

\item{var.max}{Maximum number of variables to include. Defaults to all, \code{var.max=ncol(var.mat)}.}
}
\description{
The \code{bioEnv} function performs Clarke and Ainsworth's (1993) "BIO-ENV" routine which compares (via a Mantel test)
a fixed matrix of similarities to a variable one that test all possible variable combinations.
}
\details{
The R package "vegan" contains a version of Clarke and Ainsworth's (1993)
BIOENV analysis allowing for the comparison of distance/similarity matrices between
two sets of data having either samples or variables in common.
The typical setup is in the exploration of environmental variables
that best correlate to sample similarities of the biological community
(e.g. species biomass or abundance).
In this case, the similarity matrix of the community is fixed, while subsets of
the environmental variables are used in the calculation of the environmental similarity matrix.
A correlation coefficient (typically Spearman rank correlation coefficient) is then
calculated between the two matrices and the best subset of environmental variables
can then be identified and further subjected to a permutation test to determine significance.
Due to the inflexibility of the bioenv() function, one has little control over how the variable similarity matrix is calculated (derived from the environmental subsets in the above example) as the method assumes the subset data to be environmental and that the resulting similarity matrix should be based on normalized "euclidean" distances. This makes sense with environmental data where one normalizes the data to remove the effect of differing units between parameters, yet in cases where the variable matrix is biological one might want more flexibility (a Bray-Curtis measure of similarity is common given its non-parametric nature). The vegan function vegdist() comes with many other possible indices that could be applied ("manhattan", "euclidean", "canberra", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial" and "chao"). For example, beyond the typical biological to environmental comparison (BIOENV setup), one can also use the routine to explore other other types of relationships; e.g.:
- ENVBIO: subset of biological variables that best correlate to the overall environmental pattern
- BIOBIO: subset of biological variables that best correlate to the overall biological pattern
- ENVENV: subset of environmental variables that best correlate to the overall environmental pattern

It is important to mention here that one of the reasons why a variable biological similarity
matrix is often less explored with the routine is that the number of possible subset combinations
becomes computationally overwhelming when the number of species/groups is large - the total
number of combinations being equal to 2^n - 1, where n is the total number of variables.
For this reason, Clarke and Warwick (1998) presented a stepwise routine (BVSTEP) for a faster
exploration of the subset combinations.
They specifically looked at the BIOBIO type exploration and addressed the concept of
structural redundancy in community composition through the identification of "response units",
or Taxonomic/functional groupings of species that changed in abundance in the same way over time.
}
\examples{
\donttest{
library(vegan)
data(varespec)
data(varechem)

res <- bioEnv(wisconsin(varespec), varechem,
              fix.dist.method="bray", var.dist.method="euclidean",
              scale.fix=FALSE, scale.var=TRUE
)
res
}
}

