% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cov4gappy.R
\name{cov4gappy}
\alias{cov4gappy}
\title{Covariance matrix calculation for gappy data}
\usage{
cov4gappy(F1, F2 = NULL)
}
\arguments{
\item{F1}{A data field.}

\item{F2}{An optional 2nd data field.}
}
\value{
A matrix with covariances between columns of \code{F1}.
If both \code{F1} and \code{F2} are provided, then the covariances
between columns of \code{F1} and the columns of \code{F2} are returned.
}
\description{
This function calculates a covoriance matrix for data that contain
missing values ('gappy data').
}
\details{
This function gives comparable results to \code{cov(F1, y=F2, use="pairwise.complete.obs")}
whereby each covariance value is divided by n number of shared values (as opposed
to n-1 in the case of \code{cov()}. Futhermore, the function will return a 0 (zero) in
cases where no shared values exist between columns; the advantage being that a
covariance matrix will still be calculated in cases of very gappy data, or when
spatial locations have accidentally been included without observations (i.e. land
in fields of aquatic-related parameters).
}
\examples{
# Create synthetic data
set.seed(1)
mat <- matrix(rnorm(500, sd=10), nrow=50, ncol=10)
matg <- mat
matg[sample(length(mat), 0.5*length(mat))] <- NaN # Makes 50\% missing values
matg # gappy matrix

# Calculate covariance matrix and compare to 'cov' function output
c1 <- cov4gappy(matg)
c2 <- cov(matg, use="pairwise.complete.obs")
plot(c1,c2, main="covariance comparison", xlab="cov4gappy", ylab="cov")
abline(0,1,col=8)
}
\keyword{EOF}
\keyword{PCA}
\keyword{covariance}
\keyword{gappy}

