% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2yeardec.R
\name{date2yeardec}
\alias{date2yeardec}
\title{Date - Year conversion}
\usage{
date2yeardec(date)
}
\arguments{
\item{date}{a date (class 'Date')}
}
\value{
a scalar (class 'numeric')
}
\description{
Convert dates to numeric years with decimal as fraction of a year
}
\examples{
date2yeardec(Sys.Date())

}
