% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/earthBear.R
\name{earthBear}
\alias{earthBear}
\title{Directional bearing between two geographic locations}
\usage{
earthBear(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{longitude 1 (in decimal degrees)}

\item{lat1}{Latitude 1 (in decimal degrees)}

\item{lon2}{longitude 2 (in decimal degrees)}

\item{lat2}{Latitude 2 (in decimal degrees)}
}
\value{
Vector of directional bearings (degrees)
}
\description{
\code{earthBear} calculates bearing (in degrees) between two
lon/lat positions.
One of the lon/lat positions (i.e. lon1 and lat1) can be a vector
of positions to compare against the other lon/lat position (i.e. lon2 and lat2)
}
\examples{
earthBear(0,0,20,20)
}
\keyword{geographic}

