% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/earthDist.R
\name{earthDist}
\alias{earthDist}
\title{Earth distance between two geographic locations}
\usage{
earthDist(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{longitude 1 (in decimal degrees)}

\item{lat1}{Latitude 1 (in decimal degrees)}

\item{lon2}{longitude 2 (in decimal degrees)}

\item{lat2}{Latitude 2 (in decimal degrees)}
}
\value{
Vector of distances (km)
}
\description{
\code{earthDist} calculates distance (in kilometers) between two
lon/lat positions.
The function assumes a mean equatorial Earth radius of 6378.145 km. One of the
lon/lat positions (i.e. lon1 and lat1) can be a vector of positions to compare against
the other lon/lat position (i.e. lon2 and lat2)
}
\examples{
earthDist(0,0,20,20)
}
\keyword{geographic}

