% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eof.R
\name{eof}
\alias{eof}
\title{EOF (Empirical Orthogonal Functions analysis)}
\usage{
eof(
  F1,
  center = TRUE,
  scale = FALSE,
  nu = NULL,
  method = NULL,
  recursive = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{F1}{A data field. The data should be arraunged as samples in the column 
dimension (typically each column is a time series for a spatial location).}

\item{center}{Logical (\code{TRUE/FALSE}) to define if \code{F1} should be 
centered prior to the analysis. Defaults to \code{center = TRUE}}

\item{scale}{Logical (\code{TRUE/FALSE}) to define if \code{F1} should be 
scaled prior to the analysis. Defaults to \code{scale = FALSE}}

\item{nu}{Numeric value. Defines the number of EOFs to return. Defaults to 
return the full set of EOFs.}

\item{method}{Method for matrix decomposition (\code{\link[base]{svd}}, 
\code{\link[base]{eigen}}, \code{\link[irlba]{irlba}}, 
\code{\link[RSpectra]{svds}}). 
Defaults to \code{"svd"} when \code{method = NULL} and \code{"svds"} when 
\code{method = NULL} and \code{recursive = TRUE}.
Use of \code{"svds"} or \code{"irlba"} calculates a partial SVD, which is 
recommended when \code{recursive = TRUE} due to faster computation speed.}

\item{recursive}{logical. When \code{TRUE}, the function follows the method of
"Recursively Subtracted Empirical Orthogonal Functions" (RSEOF) (Taylor et al. 2013). 
RSEOF is a modification of a least squares EOF approach for gappy data (LSEOF)
(see von Storch and Zwiers 1999)}

\item{verbose}{logical. Should progress be printed 
(when \code{"recursive = TRUE"}, Default: verbose = TRUE).}
}
\value{
Results of \code{eof} are returned as a list containing the following components:
\tabular{rll}{
\tab \code{u} \tab EOFs.\cr
\tab \code{Lambda} \tab Singular values.\cr
\tab \code{A} \tab EOF coefficients (i.e. 'Principal Components').\cr
\tab \code{tot.var} \tab Total variance of field \code{F1} (from \code{cov4gappy}). \cr
\tab \code{F1_dim} \tab Dimensions of field \code{F1}.\cr
\tab \code{F1_center} \tab Vector of center values from each column in field \code{F1}.\cr
\tab \code{F1_scale} \tab Vector of scale values from each column in field \code{F1}.\cr
}
}
\description{
This function conducts an Empirical Orthogonal Function (EOF) analysis  
via a covariance matrix (\code{cov4gappy} function) and is especially designed 
to handle gappy data (i.e. containing missing values - NaN)
}
\details{
Taylor et al. (2013) demonstrated that the RSEOF approach 
(i.e. \code{recursive = TRUE}) more accurately estimates EOFs from a 
gappy field than the traditional LSEOF method. 
Pre-treatment of gappy fields through in EOF interpolation 
(\code{\link[sinkr]{dineof}}) may provide the most accurate estimate of EOFs; 
however, RSEOF can be much faster in cases where the number of columns in \code{F1} 
is smaller than the number of rows.

All methods should give identical results when \code{recursive = TRUE}, despite 
differences in calculation time. 
When \code{recursive = FALSE}, \code{"svd"} and \code{"eigen"} give similar 
results for non-gappy fields, but will differ with gappy fields 
due to decomposition of a nonpositive definite covariance matrix, as calculated 
with \code{\link[sinkr]{cov4gappy}}. 
Specifically, \code{"eigen"} will produce negative eigenvalues for trailing 
EOFs, while singular values derived from \code{"svd"} will be strictly positive.
Faster computation time with \code{"svds"} over 
\code{"irlba"} may not result when  
\code{recursive = TRUE} due to the iterative computation of leading vectors only.
}
\examples{

# EOF of 'iris' dataset
Et <- eof(iris[,1:4])
plot(Et$A, col=iris$Species)

# Compare to results of 'prcomp'
Pt <- prcomp(iris[,1:4])
SIGN <- diag(sign(diag(cor(Et$A, Pt$x)))) # correction for differing sign
matplot(Et$A \%*\% SIGN, Pt$x)
abline(0,1, col=8)

# Compare to a gappy dataset (sign of loadings may differ between methods)
iris.gappy <- as.matrix(iris[,1:4])
set.seed(1)
iris.gappy[sample(length(iris.gappy), 0.25*length(iris.gappy))] <- NaN
Eg <- eof(iris.gappy, method="svd", recursive=TRUE) # recursive ("RSEOF")
op <- par(mfrow=c(1,2))
plot(Et$A, col=iris$Species)
plot(Eg$A, col=iris$Species)
par(op)



### EOF of climate field example
library(maps) # required package to add map

# load data
data(sst)
names(sst)

# EOF
E <- eof(sst$field, nu=10)

# Plot of EOF loading and PC
eof.num <- 2 # EOF number to plot
op <- par(no.readonly=TRUE)
layout(matrix(c(1,3,2,3),nrow=2, ncol=2), widths=c(5,1), heights=c(3,3), respect=TRUE)
op <- par(ps=10, cex=1)
par(mar=c(4,4,1,1))
PAL <- colorPalette(c("blue", "cyan", "grey90", "yellow", "red"), c(10,1,1,10))
ZLIM <- c(-1,1)*max(abs(E$u[,eof.num]))
COL <- val2col(E$u[,eof.num], col=PAL(100), zlim=ZLIM)
plot(lat ~ lon, data=sst$grid, pch=22, bg=COL, col=COL, cex=2)
map("world", add=TRUE)
par(mar=c(4,0,1,4))
imageScale(E$u[,eof.num], col=PAL(100), zlim=ZLIM, axis.pos=4)
par(mar=c(4,4,1,4))
plot(sst$date, E$A[,eof.num], t="l", xlab="date", ylab="")
lines(loess.smooth(sst$date, E$A[,eof.num], span=1/3), col=rgb(0.5,0.5,1), lwd=2) # smoothed signal
abline(h=0, v=seq(as.Date("1000-01-01"), as.Date("2100-01-01"), by="10 years"), col=8, lty=3)
par(op)


}
\references{
Bjoernsson, H. and Venegas, S.A. (1997). "A manual for EOF and SVD 
analyses of climate data", McGill University, CCGCR Report No. 97-1, 
Montreal, Quebec, 52pp.

von Storch, H, Zwiers, F.W. (1999). Statistical analysis in climate 
research. Cambridge University Press.

Taylor, Marc H., Martin Losch, Manfred Wenzel, Jens Schroeter (2013). 
On the Sensitivity of Field Reconstruction and Prediction Using 
Empirical Orthogonal Functions Derived from Gappy Data. J. Climate, 
26, 9194-9205. \href{http://dx.doi.org/10.6084/m9.figshare.732650}{pdf}
}
\keyword{EOF}
\keyword{PCA}
\keyword{gappy}
