% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eofBoot.R
\name{eofBoot}
\alias{eofBoot}
\title{Calculate number of non-mixed EOFs (eof version)}
\usage{
eofBoot(
  F1,
  center = TRUE,
  scale = FALSE,
  nu = NULL,
  method = NULL,
  recursive = FALSE,
  nperm = 99,
  verbose = TRUE
)
}
\arguments{
\item{F1}{A data field. The data should be arraunged as samples in the column 
dimension (typically each column is a time series for a spatial location).}

\item{center}{Logical (\code{TRUE/FALSE}) to define if \code{F1} should be 
centered prior to the analysis. Defaults to 'TRUE'}

\item{scale}{Logical (\code{TRUE/FALSE}) to define if \code{F1} should be 
scaled prior to the analysis. Defaults to 'TRUE'}

\item{nu}{Numeric value. Defines the number of EOFs to return. Defaults to 
return the full set of EOFs.}

\item{method}{Method for matrix decomposition ('\code{svd}', '\code{eigen}', 
'\code{irlba}'). Defaults to 'svd' when \code{method = NULL}. Use of 'irlba' can 
  dramatically speed up computation time when \code{recursive = TRUE} but may
  produce errors in computing trailing EOFs. Therefore, this option is only 
  advisable when the field \code{F1} is large and when only a partial 
  decomposition is desired (i.e. \code{nu << dim(F1)[2]}). 
  All methods should give identical results when \code{recursive=TRUE}.
\code{svd} and \code{eigen} give similar results for non-gappy fields, 
  but will differ slightly with gappy fields due to decomposition of a 
  non-positive definite covariance matrix. Specifically, \code{eigen}  will 
  produce negative eigenvalues for trailing EOFs, while singular values 
  derived from \code{svd} be strictly positive.}

\item{recursive}{Logical. When \code{TRUE}, the function follows the method 
of "Recursively Subtracted Empirical Orthogonal Functions" (RSEOF). 
See \code{\link[sinkr]{eof}} for details.}

\item{nperm}{Numeric. The number of null model permutations to calculate.}

\item{verbose}{logical. Print progress (Default: verbose = TRUE).}
}
\description{
The \code{eofBoot} function uses a bootstrap randomization approach to 
  calculate distributions of Empirical Orthogonal Function analysis (EOF) 
  singular values with the \code{\link[sinkr]{eof}} function.
  EOF mode significance is assessed against the distributions of 
  neighboring EOF singular values ("Lambda") calculated by the permutated 
  models. A bootstrap routine follows the procedure of 
  Babamoradi et al. (2013) whereby permutations sample rows (samples) more 
  than once, which is a non-parametric approach does not make assumptions 
  about the distribution of data.
}
\examples{
data(Xt)

# The "noisy" field
# noise standard deviation at 10\% noise-to-signal ratio
noise_sd <- 0.1 * sd(as.vector(Xt))

# Add Gaussian noise
set.seed(123)  # For reproducibility
Xn <- Xt + rnorm(length(Xt), mean = 0, sd = noise_sd)
Xn <- array(Xn, dim = dim(Xt))

res <- eofBoot(Xn, center = FALSE, scale = FALSE, nperm=499)
ylim <- range(res$Lambda.orig, res$Lambda)
boxplot(res$Lambda, log="y", col=8, border=2, outpch="", ylim=ylim)
points(res$Lambda.orig)
abline(v=res$n.sig+0.5, lty=2, col=4)
mtext(paste("Non-mixed PCs =", res$n.sig), side=3, line=0.5, col=4)


}
\references{
Babamoradi, H., van den Berg, F., Rinnan, A, 2013. Bootstrap based 
confidence limits in principal component analysis - A case study, 
Chemometrics and Intelligent Laboratory Systems, Volume 120,
pp. 97-105. doi:10.1016/j.chemolab.2012.10.007.
}
