% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResponse.R
\name{getResponse}
\alias{getResponse}
\title{Return response variable name from a fitted model}
\usage{
getResponse(formula)
}
\arguments{
\item{formula}{or fitted model object}
}
\value{
a vector of response variable names as character strings
}
\description{
Return response variable name from a fitted model
}
\details{
Works for lm, glm, gam and gls model objects - maybe others as well.
}
\examples{
getResponse(formula(y ~ x)) # single response variable
getResponse(formula(y1 + y2 ~ x)) # multiple response variables

x <- runif(10)
y <- 2*x + rnorm(10)
fit <- lm(y ~ x)
getResponse(fit) # using a fitted lm model

fit <- glm(y ~ x)
getResponse(fit) # using a fitted glm model

}
