% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isin.convert.R
\name{isin.convert}
\alias{isin.convert}
\title{Calculate grid and polygon coordinates for 4 km ISIN grid}
\usage{
isin.convert(grd = NULL, coord = NULL, polygons = FALSE)
}
\arguments{
\item{grd}{numeric vector containing ISIN grid numbers needed for calculating 
lon/lat coordinates or polygons.}

\item{coord}{data.frame containing ISIN grid coordinates (row and column 
locations). Names of columns should be \code{coord$row} and \code{coord$col}.}

\item{polygons}{Logical. Should grid polygon corner coordinates be returned.}
}
\value{
If "\code{polygons = TRUE}", then returns grid information 
as a dataframe, including lon / lat values of grid centers. If 
"\code{polygons = TRUE}", then returns a list with
polygon corners in a dataframe(longitudinal coordinates of corners (e.g. 
\code{[[i]]$x}) and latitudinal coordinates of corners ( 
\code{[[i]]$y}))
}
\description{
This function is used to convert information regarding the 
ISIN grid information used by Globcolour as well as to construct 
associated polygons for use in mapping. 
The raw Globcolour .nc files come with column and row pointers 
as to the the grid's location. For 4.63 km resolution data ("L3b"), this
translates to 4320 latitudinal rows with varying number of 
associated longitudinal columns depending on the latitudinal
circumference. 
Input must be either a vector of grid numbers, \code{grd}, or a dataframe
with column and row identifiers, \code{coord}. See Globcolour's "Product 
User Guide" for fulther details of the ISIN grid 
(\url{http://www.globcolour.info/products_description.html}).
}

