% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{jetPal}
\alias{jetPal}
\title{jet palette}
\usage{
jetPal(n)
}
\description{
\code{jetPal} is a palette of colors similar that found in Matlab.
This is a nice general palette for data exploration - similar to a rainbow palette
but does not cycle back to the lower color level
(dark blue --> cyan --> yellow --> dark red).
}
\examples{
image(volcano, col=jetPal(50))
}

