% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lonLatFilter.R
\name{lonLatFilter}
\alias{lonLatFilter}
\title{Filter lon/lat positions that fall within defined boundaries}
\usage{
lonLatFilter(lon_vector, lat_vector, west, east, south, north)
}
\arguments{
\item{lon_vector}{Longitude 1 (in decimal degrees)}

\item{lat_vector}{Latitude 1 (in decimal degrees)}

\item{west}{West longitude border (decimal degrees)}

\item{east}{East longitude border (decimal degrees)}

\item{south}{South latitude border (decimal degrees)}

\item{north}{North latitude border (decimal degrees)}
}
\value{
Vector of position inclusion in the defined lon/lat borders
}
\description{
\code{lonLatFilter} Tests whether lon/lat positions fall within a defined
box of lon/lat borders (location on border returns \code{TRUE}
}
\examples{
set.seed(1)
n <- 1000
Pos <- list(lon=runif(n, min=-180, max=180), lat=runif(n, min=-180, max=180))
# Check to see if positions are withing boundaries
res <- lonLatFilter(Pos$lon, Pos$lat, -20, 20, -40, 40)
# Plot
op <- par(mar=c(4,4,1,1))
plot(Pos$lon, Pos$lat, pch=21, col=1, bg=2*res)
rect(-20, -40, 20, 40, border=3)
par(op)
}
\keyword{geographic}

