% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lsos.R
\name{lsos}
\alias{lsos}
\title{List top n objects by size}
\usage{
lsos(pos = 1, pattern, order.by = "Size", decreasing = TRUE,
  head = TRUE, n = 10)
}
\arguments{
\item{pos}{where to look for the object (see \code{\link[base]{get}}, 'Details');
if omitted search as if the name of the object appeared unquoted in an expression.}

\item{pattern}{an optional regular expression (see \code{\link[base]{ls}}). Only names matching pattern are returned.
see \code{\link[utils]{glob2rx}} can be used to convert wildcard patterns to regular expressions.}

\item{order.by}{Order list by one of the following: \code{"Type", "Size", "Rows", "Columns"}}

\item{decreasing}{Logical. Decreasing in size (i.e. largest objects at top of list.)
(Default=\code{TRUE})}

\item{head}{Logical (Default=\code{TRUE}). Should only the \code{\link[utils]{head}} be
returned in the results of a \code{data.frame}. Argument \code{n} defines how many objects to include.}

\item{n}{Numeric value (Default=\code{10}). Defines how many objects to include in results}
}
\value{
a \code{data.frame} object
}
\description{
The function \code{lsos} lists the top \code{n} objects in memory.
}
\examples{
x1 <- matrix(rnorm(10000), 100, 100)
x2 <- as.data.frame(matrix(rnorm(1000), 100, 10))
x3 <- rnorm(1000)
lsos()
rm(list=c("x1", "x2", "x3"))
}
\references{
\url{http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
}

