% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maga.R, R/mantelGA.R
\name{magaBest}
\alias{magaBest}
\title{Extract top models from an mantelGA object}
\usage{
magaBest(fit, output.best = 10)

magaBest(fit, output.best = 10)
}
\arguments{
\item{fit}{output from maga function}

\item{output.best}{Number of best combinations to return in the results 
object (Default `output.best = 10`).}
}
\value{
list containing `order.by.best`, `order.by.i.comb`, 
  `best.model.vars`, `best.model.rho`

list containing `order.by.best`, `order.by.i.comb`, 
  `best.model.vars`, `best.model.rho`
}
\description{
Extract top models from an mantelGA object

Extract top models from an mantelGA object
}
\examples{




}
