% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maga.R
\name{magaMonitor}
\alias{magaMonitor}
\title{Output Genetic Algorithm population fitness evolution}
\usage{
magaMonitor(object, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{an object of class ga-class or gaisl-class, usually resulting 
from a call to function ga or gaisl, respectively.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\value{

}
\description{
Called internally by `GA::ga`, and used within 
  `maga`.
}
