% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGlobcolourField.R
\name{makeGlobcolourField}
\alias{makeGlobcolourField}
\title{Make a field from Globcolour data}
\usage{
makeGlobcolourField(dataPath = getwd(), yearRan = c(1997, 2013),
  lonRan = c(-83, -79), latRan = c(-16, -4), greps = c(".nc", "GSM",
  "MO"))
}
\arguments{
\item{dataPath}{Path to data}

\item{yearRan}{Range of years to use for field}

\item{lonRan}{Range of longitude to use for field}

\item{latRan}{Range of latitude to use for field}

\item{greps}{Text patterns to use to filter data (i.e. filter file names
that only contain these text strings)}
}
\value{
Matrix field where rows and columns are time and space, respectively
}
\description{
Make a field from Globcolour data
}

