% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixIndex.R
\name{matrixIndex}
\alias{matrixIndex}
\title{Convert matrix element index to row/column coordinate or vice-versa}
\usage{
matrixIndex(idx = NULL, coord = NULL, dim.mat = NULL)
}
\arguments{
\item{idx}{numeric. Index of a matrix element}

\item{coord}{numeric vector or matrix. Coordinate of an index (row,col)}

\item{dim.mat}{numeric vector. Dimensions of matrix (rows, ncols)}
}
\value{
either a vector (indices) or matrix (coordinate) element
}
\description{
Convert matrix element index to row/column coordinate or vice-versa
}
\examples{
M <- matrix(1:20, 4,5)
M
matrixIndex(idx = 15, dim.mat = dim(M))
matrixIndex(coord = c(3,4), dim.mat = dim(M))
matrixIndex(coord = matrix(c(3,4,4,5), ncol=2, byrow=TRUE), 
  dim.mat = dim(M))


}
