% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrixPoly.R
\name{matrixPoly}
\alias{matrixPoly}
\title{Make polygons from a matrix}
\usage{
matrixPoly(x, y, z, n = NULL)
}
\arguments{
\item{x,y}{Optional vectors of values for matrix \code{z} rows (\code{x}) and columns
(\code{y}). If excluded, the function assumes the values to be a evenly-spaced
sequence from 0 to 1.}

\item{z}{A matrix}

\item{n}{An optional vector of element positions of \code{z} for polygon creation}
}
\value{
List of polygon coordinates for each element of \code{z}
}
\description{
\code{matrixPoly} creates a list of polygon coordinates given a matrix \code{z} and
corresponding \code{x} and \code{y} coordinates for the dimensions of \code{z}.
}
\examples{
# Make sythetic data
set.seed(1)
m=8
n=10
x <- seq(m)
y <- seq(n)
z <- matrix(runif(m*n), nrow=m, ncol=n)

# Ex 1 - add another image layer
image(x, y, z, col=grey.colors(20))
N <- sample(1:(m*n),20)
z2 <- NaN*z
z2[N] <- 1
image(x, y, z2, col=rgb(0,0,1,0.4), add=TRUE)
box()

# Ex 2 - add polygons
image(x, y, z, col=grey.colors(20))
poly <- matrixPoly(x, y, z=z, n=N)
sapply(poly, function(X){polygon(X, col=rgb(1,1,0,0.4), border=1)})
box()

# Ex 3 - add polygons to unequal grid
x2 <- cumsum(round(runif(m, min=1, max=10)))
y2 <- cumsum(round(runif(n, min=1, max=10)))
image(x2, y2, z, col=grey.colors(20))
poly <- matrixPoly(x2, y2, z=z, n=N)
sapply(poly, function(X){polygon(X, col=rgb(1,0,0,0.4), border=1)})
box()
}

