% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nearest.R
\name{nearest}
\alias{nearest}
\title{Calculate the nearest element in a vector as compared to a reference value}
\usage{
nearest(value, lookup_vector)
}
\arguments{
\item{value}{A numeric reference value}

\item{lookup_vector}{The vector to compare to the reference value}
}
\value{
Vector element index of nearest value
}
\description{
the \code{nearest} function returns the position of a vector that is closest to a defined
value by determining the element with the smallest squared distance.
}
\examples{
set.seed(1)
x <- runif(10, min=0, max=100)
res <- nearest(50, x)
plot(x)
abline(h=50, col=8, lty=2)
points(res, x[res], pch=20, col=2)
}

