% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/newRange.R
\name{newRange}
\alias{newRange}
\title{Define a new range for a numeric}
\usage{
newRange(x, new.min = 0, new.max = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{new.min}{numeric value of new minimum range value}

\item{new.max}{numeric value of new maximum range value}
}
\description{
The \code{newRange} function scales a numeric vector to a new range as defined by
minimum and maximum extreme values.
}
\examples{
y <- runif(100, min=0, max=2)
y2 <- newRange(y, new.min=0, new.max=1)
plot(y, t="l")
lines(y2, col=2)

plot(y, y2)
}

