% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northTest.R
\name{northTest}
\alias{northTest}
\title{North's Rule of Thumb for EOF significance}
\usage{
northTest(x, Lambda)
}
\arguments{
\item{x}{Matrix. Field used during EOF decomposition}

\item{Lambda}{Vector of singular values representing variance magnitudes 
explained by each EOF mode. Results of \code{\link[stats]{prcomp}} will 
need to be squared from their standard deviation values 
(e.g. \code{Lambda=res$sdev^2}).}
}
\description{
The \code{northTest} function assesses the uniqueness of EOF modes through
  assumptions of error on singular values ("Lambda") as described by 
  North et al (1982). Overlapping error limits between 
  neighbouring Lambda values indicates a possible mixture of signals. 
  A similar test via a bootstrapping approach (possibly more robust) 
  can be done with the \code{\link[sinkr]{prcompBoot}} or 
  \code{\link[sinkr]{eofBoot}} functions.
}
\examples{
data(Xt)

# The "noisy" field
# noise standard deviation at 10\% noise-to-signal ratio
noise_sd <- 0.1 * sd(as.vector(Xt))

# Add Gaussian noise
set.seed(123)  # For reproducibility
Xn <- Xt + rnorm(length(Xt), mean = 0, sd = noise_sd)
Xn <- array(Xn, dim = dim(Xt))

# eof + northTest 
E <- eof(Xn, center=FALSE, scale=FALSE)
L <- E$Lambda # Lambdas
res <- northTest(Xn, L)
plot(L, ylim=range(c(L+res$upper.lim, L-res$lower.lim)), log="y")
segments(seq(L), L-res$lower.lim, seq(L), L+res$upper.lim, col=2)
abline(v=res$n.sig+0.5, lty=2, col=4)
mtext(paste("Non-mixed PCs =", res$n.sig), side=3, line=0.5, col=4)

# prcomp + northTest 
E <- prcomp(Xn, center=FALSE, scale=FALSE)
L <- E$sdev^2 # Lambdas
res <- northTest(Xn, L)
plot(L, ylim=range(c(L+res$upper.lim, L-res$lower.lim)), log="y")
segments(seq(L), L-res$lower.lim, seq(L), L+res$upper.lim, col=2)
abline(v=res$n.sig+0.5, lty=2, col=4)
mtext(paste("Non-mixed PCs =", res$n.sig), side=3, line=0.5, col=4)


  
}
\references{
G.R. North, T.L. Bell, R.F. Cahalan, and F.J. Moeng. (1982). 
Sampling errors in the estimation of empirical orthogonal functions. 
Mon. Wea. Rev., 110:699-706.
}
