% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maga.R, R/mantelGA.R
\name{plot.magaBest}
\alias{plot.magaBest}
\title{plot results from magaBest}
\usage{
\method{plot}{magaBest}(res)

\method{plot}{magaBest}(res)
}
\arguments{
\item{res}{results of call to `magaBest`}
}
\value{
a plot

a plot
}
\description{
plot results from magaBest

plot results from magaBest
}
\examples{








}
