% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegionCol.R
\name{plotRegionCol}
\alias{plotRegionCol}
\title{Add color to plot region}
\usage{
plotRegionCol(col = "grey90", grid = FALSE, grid.col = "white", ...)
}
\arguments{
\item{col}{Color string or number. Color of plot region (Defaut: "grey90")}

\item{grid}{Logical. Add grid on top of plot region color (Default: FALSE)}

\item{grid.col}{color of grid lines (Default: "white")}

\item{...}{Other arguments to pass to \code{\link[graphics]{grid}}}
}
\value{
adds color to plot region and optional grid
}
\description{
Add color to plot region
}
\examples{
plot(dist~speed, data=cars)
plotRegionCol()
points(dist~speed, data=cars, pch=21, bg="white")

# with grid
plot(dist~speed, data=cars)
plotRegionCol(grid=TRUE)
points(dist~speed, data=cars, pch=21, bg="cyan")

# with grid plus addition settings
plot(dist~speed, data=cars)
plotRegionCol(col="grey30", grid=TRUE, grid.col="grey50", lwd=0.5)
points(dist~speed, data=cars, pch=21, bg="white")

}
