% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcompBoot.R
\name{prcompBoot}
\alias{prcompBoot}
\title{Calculate number of non-mixed EOFs (prcomp version)}
\usage{
prcompBoot(
  x,
  retx = TRUE,
  center = TRUE,
  scale. = FALSE,
  tol = NULL,
  nperm = 99
)
}
\arguments{
\item{x, retx, center, scale., tol}{See \code{\link[stats]{prcomp}} 
for argument definitions.}

\item{nperm}{Numeric. The number of null model permutations to calculate.}
}
\description{
The \code{prcompBoot} function uses a bootstrap randomization approach to 
calculate distributions of Empirical Orthogonal Function analysis (EOF) 
singular values with the \code{\link[stats]{prcomp}} function.
EOF mode significance is assessed against the distributions of 
neighboring EOF singular values ("Lambda") calculated by the permutated 
models. A bootstrap routine follows the procedure of Babamoradi et al. (2013) whereby 
permutations sample rows (samples) more than once, which is a non-parametric 
approach does not make assumptions about the distribution of data.
}
\examples{
data(Xt)

# The "noisy" field
# noise standard deviation at 10\% noise-to-signal ratio
noise_sd <- 0.1 * sd(as.vector(Xt))

# Add Gaussian noise
set.seed(123)  # For reproducibility
Xn <- Xt + rnorm(length(Xt), mean = 0, sd = noise_sd)
Xn <- array(Xn, dim = dim(Xt))

res <- prcompBoot(Xn, center=FALSE, scale=FALSE, nperm=499)
ylim <- range(res$Lambda.orig, res$Lambda)
boxplot(res$Lambda, log="y", col=8, border=2, outpch="", ylim=ylim)
points(res$Lambda.orig)
abline(v=res$n.sig+0.5, lty=2, col=4)
mtext(paste("Non-mixed PCs =", res$n.sig), side=3, line=0.5, col=4)


}
\references{
Babamoradi, H., van den Berg, F., Rinnan, A, 2013. Bootstrap based 
confidence limits in principal component analysis - A case study, 
Chemometrics and Intelligent Laboratory Systems, Volume 120,
pp. 97-105. doi:10.1016/j.chemolab.2012.10.007.
}
