% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcompRecon.R
\name{prcompRecon}
\alias{prcompRecon}
\title{prcomp object reconstruction}
\usage{
prcompRecon(pca, pcs = NULL, unscale = TRUE, uncenter = TRUE)
}
\arguments{
\item{pca}{An object resulting from the function \code{\link[stats]{prcomp}}.}

\item{pcs}{The principal components ("PCs") to use in the reconstruction 
(defaults to the full set of PCs: \code{pcs=seq(pca$sdev))})}

\item{unscale}{logical. Should reconstruction be unscaled 
(reverses scaling; Default: `unscale = TRUE`).}

\item{uncenter}{logical. Should reconstruction be uncentered 
(reverses centering; Default: `uncenter = TRUE`).}
}
\description{
This function reconstructs the original field from an EOF object of the 
function \code{\link[stats]{prcomp}}.
}
\examples{
# prcomp
P <- prcomp(iris[,1:4])

# Full reconstruction
R <- prcompRecon(P)
plot(as.matrix(iris[,1:4]), R, xlab = "original data", 
  ylab = "reconstructed data")
abline(0, 1, col=2)

# Partial reconstruction
RMSE <- NaN*seq(P$sdev)
for(i in seq(RMSE)){
  Ri <- prcompRecon(P, pcs=seq(i))
  RMSE[i] <- sqrt(mean((as.matrix(iris[,1:4]) - Ri)^2))
}
plot(RMSE, t="o", xlab="Number of pcs")
abline(h=0, lty=2)

}
