% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ptlocator.R
\name{ptlocator}
\alias{ptlocator}
\title{Data point locator}
\usage{
ptlocator(n = 1, x, y, col = rgb(1, 0, 0, 0.5), pch = 20, ...)
}
\arguments{
\item{n}{Number of points to select}

\item{x}{Vector of x-axis values for the plotted data}

\item{y}{Vector of y-axis values for the plotted data}

\item{col}{Colors to use for plotting closest data points}

\item{pch}{pch values to use for plotting closest data points}

\item{...}{additional parameters for plotting closest data points}
}
\value{
A vector of point indices
}
\description{
the \code{ptlocator} function allows for the selection of data points in
an open graphical device via the \code{\link[graphics]{locator}} function. Closest data points
to the selected positions are determined via Euclidean distance following a scaling of x and y
axes in order to give them equal weighting and remove the influence of differing units or ranges.
Colored points are filled in for the data point that has the lowest distance to the
clicked location, and the result gives the vector positions of the closest x, y data points.
}
\examples{
\donttest{set.seed(1)
n <- 200
x <- sort(runif(n, min=0, max=10*pi))
y <- sin(x) + rnorm(n, sd=0.2)

# Select 10 points at maxima and minima
plot(x, y)
pos <- ptlocator(10, x, y)
pos}
}

