% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate a matrix by a quarter-turn clockwise}
\usage{
rotate(mat)
}
\arguments{
\item{mat}{matrix to be rotated.}
}
\value{
matrix
}
\description{
Rotate a matrix by a quarter-turn clockwise
}
\examples{
op <- par(mfcol = c(1,2))
image(volcano)
image(rotate(volcano))
par(op)

}
