% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_along.R
\name{scale_along}
\alias{scale_along}
\title{Scale an array along a single dimension}
\usage{
scale_along(arr, dim_to_scale = 3, center = TRUE, scale = TRUE)
}
\arguments{
\item{arr}{array. Object to scale.}

\item{dim_to_scale}{numeric. Dimension}

\item{center}{logical. Passed to \code{\link[base]{scale}}}

\item{scale}{logical. Passed to \code{\link[base]{scale}}}
}
\value{
array
}
\description{
Scale an array along a single dimension
}
\examples{

set.seed(1111)
A <- array(rnorm(24), dim = c(2,3,4), 
  dimnames = list(
    paste("a", seq(2), sep = "~"),
    paste("b", seq(3), sep = "~"),
    paste("c", seq(4), sep = "~")
  )
)
A
Asc <- scale_along(A, 3)
Asc
apply(Asc, 1:2, mean) # mean = 0.0
apply(Asc, 1:2, sd) # sd = 1.0

}
