% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqRan.R
\name{seqRan}
\alias{seqRan}
\title{Create a sequence of defined length over range of a vector}
\usage{
seqRan(x, length.out = NULL, relExt = c(0, 0), ...)
}
\arguments{
\item{x}{Numeric vector}

\item{length.out}{Length of sequence to create}

\item{relExt}{Relative extention of sequence limits. Default \code{relExt = c(0,0)} maintains 
original range of \code{x}}

\item{...}{Additional parameters to pass to \code{\link[base]{seq}}}
}
\value{
Vector
}
\description{
Create a sequence of defined length over range of a vector
}
\examples{
x <- 1:5
seqRan(x, 10) # using original range of x
seqRan(x, 10, c(-0.1, 0.1)) # using extended range of x

}

