% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowtext.R
\name{shadowtext}
\alias{shadowtext}
\title{Add border to plotted text}
\usage{
shadowtext(
  x,
  y = NULL,
  labels,
  col = "white",
  bg = "black",
  theta = seq(0, 2 * pi, length.out = 50),
  r = 0.1,
  ...
)
}
\arguments{
\item{x, y}{numeric vectors of coordinates where the text labels should 
be written. If the length of x and y differs, the shorter one is recycled.}

\item{labels}{a character vector or expression specifying the text to be 
written. An attempt is made to coerce other language objects 
(names and calls) to expressions, and vectors and other classed objects 
to character vectors by as.character. If labels is longer than x and y, 
the coordinates are recycled to the length of labels.}

\item{col}{the color of the text.}

\item{bg}{the color of the text border}

\item{theta}{a sequence of angles be which to cycle around the text. 
(Default: \code{theta = seq(0, 2*pi, length.out = 50)})}

\item{r}{the width of the border (fraction of character width).}

\item{...}{other pars passed to \code{\link[graphics]{text}}.}
}
\value{
lower level plotting command to add text to existing device.
}
\description{
Add border to plotted text
}
\examples{

plot(c(0,1), c(0,1), type="n", lwd=20, axes=FALSE, xlab="", ylab="")

rect(xleft = 0.5, xright = 1, ybottom = 0, ytop = 1, col=1)
text(1/6, 1/6, "Test 1")
shadowtext(2/6, 2/6, "Test 2", col='red', bg="blue")
shadowtext(3/6, 3/6, "Test 3", cex=2)

# `r` controls the width of the border
shadowtext(5/6, 5/6, "Test 4", col="black", bg="white", cex=4, r=0.2)

}
\references{
\url{https://stackoverflow.com/a/25632536/1199289}
}
