% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedVals.R
\name{sharedVals}
\alias{sharedVals}
\title{Identify shared values between two vectors}
\usage{
sharedVals(x, y)
}
\arguments{
\item{x}{first vector of values}

\item{y}{second vector of values}
}
\value{
Vector of shared values
}
\description{
Identify shared values between two vectors
}
\examples{

x <- 1:5
y <- c(1,3,5,7,9)
sharedVals(x,y)

}

