% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqbin.R
\name{sqbin}
\alias{sqbin}
\title{Square binning of xy data}
\usage{
sqbin(x, y, xint = NULL, yint = NULL, nxbin = 50, nybin = 50)
}
\arguments{
\item{x}{a vector of x values}

\item{y}{a vector of y values}

\item{xint}{a vector of x intervals (end values). Defaults to values as specified
by \code{pretty(x, n=nxbin)}}

\item{yint}{a vector of y intervals (end values). Defaults to values as specified
by \code{pretty(y, n=nybin)}}

\item{nxbin}{number of x bins. Default=50. Not used when \code{xint}
is defined.}

\item{nybin}{number of y bins. Default=50. Not used when \code{yint}
is defined.}
}
\value{
A matrix containing interval mid points (\code{x,y}), bin frequencies (\code{z}), 
and intervals (\code{xint, yint}).
}
\description{
The \code{sqbin} function calculates frequencies in xy bins
}
\examples{
# Synthetic data
set.seed(1)
n <- 1e6
x <- runif(n, min=-3, max=3)
y <- 4*x^2 + rnorm(n, sd=5)

sqbin.res <- sqbin(x,y)

# Plot
op <- par(mar=c(4,4,1,1))
image(sqbin.res, col=jetPal(20))
par(op)

# Plot with legend
op <- par(no.readonly = TRUE)
lo <- matrix(1:2, nrow=1, ncol=2)
layout(lo, widths=c(4,1), heights=c(4), respect=TRUE)
par(cex=1)
par(mar=c(3,3,1,1))
image(sqbin.res, col=jetPal(20))
par(mar=c(3,0,1,3))
imageScale(sqbin.res$z, col=jetPal(20), axis.pos=4)
par(op)

}

