% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale a matrix}
\usage{
unscale(x, unscale = TRUE, uncenter = TRUE)
}
\arguments{
\item{x}{a numeric matrix(like object) that has been centered and/or scaled by the
\code{\link[base]{scale}} function.}

\item{unscale}{a logical value defining whether to unscale \code{x}.}

\item{uncenter}{a logical value defining whether to uncenter \code{x}.}
}
\description{
The \code{unscale} function unscales a numeric marteix that has been either
centered or scaled by the \code{\link[base]{scale}} function. This is done by
reversing the first unscaling and then uncentering based on the
object's attributes.
}
\examples{
x <- matrix(1:16, 4, 4)
xcs <- scale(x, center=TRUE, scale=TRUE) # centered and scaled
xc <- scale(x, center=TRUE, scale=FALSE) # centered only
xs <- scale(x, center=FALSE, scale=TRUE) # scaled only

# compare difference to original
x - unscale(xcs)
x - unscale(xc)
x - unscale(xs)
}

