% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{val2col}
\alias{val2col}
\title{Convert values to color levels}
\usage{
val2col(z, zlim, col = heat.colors(12), breaks)
}
\arguments{
\item{z}{A vector of values.}

\item{zlim}{Limits of the color scale values.}

\item{col}{Vector of color values}

\item{breaks}{Break points for color changes. If breaks is specified then \code{zlim}
is unused and the algorithm used follows \code{cut}, so intervals are
closed on the right and open on the left except for the lowest interval
which is closed at both ends.}
}
\description{
The \code{val2col} function converts a vector of values("z")
to a vector of color levels. One must define the number of colors.
The limits of the color scale ("zlim") or the break points for the
color changes("breaks") can also be defined. When breaks and zlim are
defined, breaks overrides zlim. All arguments are similar to those in the
\code{image} function.
}
\examples{
x <- 2
y <- 2
x+y
}

