% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woldcv.R
\name{woldcv}
\alias{woldcv}
\title{Wold's PCA cross-validation procedure}
\usage{
woldcv(dat, fracValid = 0.2, centered = TRUE, scaled = FALSE, ...)
}
\arguments{
\item{dat}{Dataset (matrix)}

\item{fracValid}{Fraction of data to use for each validation. Will determine the
number of cross-validation groups (Default=0.2)}

\item{centered}{Center data (logical). Passed to \link{eof}}

\item{scaled}{Scale data (logical). Passed to \link{eof}}

\item{...}{Additional parameters to pass to \link{eof}}
}
\value{
List of sum of square error (SS), root mean square error (RMSECV) of 
cross-validation, number of values used in validation (npred), number of groups
used in cross-validation (ngroups)
}
\description{
Wold's PCA cross-validation procedure
}
\examples{

data(wine)
dat <- as.matrix(wine[,-1])
res <- woldcv(dat, fracValid=0.2, centered=TRUE, scaled=TRUE)
plot(res$SS, t="b"); which.min(res$SS)
plot(res$RMSECV, t="b"); which.min(res$RMSECV)

data(iris)
dat <- as.matrix(iris[,-5])
res <- woldcv(dat, fracValid=0.3, centered=TRUE, scaled=FALSE)
plot(res$SS, t="b"); which.min(res$SS)
plot(res$RMSECV, t="b"); which.min(res$RMSECV)

}
\references{
Wold S (1978) Technometrics 20:397–405
}

