% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeardec2date.R
\name{yeardec2date}
\alias{yeardec2date}
\title{Year - Date conversion}
\usage{
yeardec2date(yeardec)
}
\arguments{
\item{yeardec}{numeric year}
}
\value{
date in format "\%Y-\%m-\%d" (class 'Date').
}
\description{
Convert numeric years to dates
}
\examples{
 yeardec2date(2014.14)

}
