% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{colorPalette}
\alias{colorPalette}
\title{Color interpolation with uneven step size}
\usage{
colorPalette(steps, n.steps.between = NULL, ...)
}
\arguments{
\item{...}{arguments to pass to \code{\link[grDevices]{colorRamp}}.}

\item{colors}{colors to interpolate; must be a valid argument to \code{col2rgb()}.}

\item{bias}{a positive number. Higher values give more widely spaced colors at the high end.}

\item{space}{a character string; interpolation in RGB or CIE Lab color spaces.}

\item{interpolate}{use spline or linear interpolation.}

\item{alpha}{logical: should alpha channel (opacity) values should be returned? It is an error to give a true value if space is specified.}
}
\description{
Color ramp with differing number of steps between color levels. Wrapper for colorRamp
}
\details{
This is a wrapper function for colorRampPalette. It allows for the
definition of the number of intermediate colors between the main colors.
Using this option one can stretch out colors that should predominate
the palette spectrum. Additional arguments of colorRampPalette can also
be added regarding the type and bias of the subsequent interpolation..
}
\examples{
op <- par(mfcol=c(2,1), omi=c(0.1,0.1,0.1,0.1), mai=c(1,0.2,0.2,0.2))
steps <- c("blue4", "cyan", "white", "yellow", "red4")
pal <- colorPalette(steps, space="rgb")
z=1:1000
imageScale(z, col=pal(41))
box()
steps <- c("blue4", "cyan", "white", "yellow", "red4")
pal <- colorPalette(steps, c(20,1,1,20), space="rgb")
z=1:1000
imageScale(z, col=pal(41))
box()
par(op)
}
\keyword{color}

