% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jetPal.R
\name{jetPal}
\alias{jetPal}
\title{jet palette}
\usage{
jetPal(n)
}
\arguments{
\item{n}{Number of colors to generate}
}
\description{
\code{jetPal} is a palette of colors similar that found in Matlab.
This is a nice general palette for data exploration - similar to a rainbow palette
but does not cycle back to the lower color level
(dark blue --> cyan --> yellow --> dark red).
}
\examples{
image(volcano, col=jetPal(50))
}

