% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_loocv.R
\name{pca_loocv}
\alias{pca_loocv}
\title{Principal component analysis "leave-one-out" cross-validation}
\usage{
pca_loocv(X, npc.max = 25)
}
\arguments{
\item{X}{Matrix to be subjected to svd}

\item{npc.max}{The maximum number of principal components to test.}
}
\value{
Matrix of square error values for each element in X
}
\description{
Principal component analysis "leave-one-out" cross-validation
}
\examples{

library(MASS)
X <- as.matrix(iris[,1:4])
res <- pca_loocv(X)
res <- lapply(res, colSums)
COL <- 2:4

op <- par(mar=c(4,4,2,1), tcl=-0.25, mgp=c(2.5,0.5,0))
for(i in seq(res)){
  if(i==1) {
    plot(res[[i]], t="n", ylim=range(unlist(res)), main="iris", xlab="n PCs", ylab="PRESS")
    grid()
  } 
  lines(res[[i]], t="b", bg=c(NaN,COL[i])[(res[[i]]==min(res[[i]])) + 1], col=COL[i], pch=21)
}
legend("topleft", legend=c("naive", "approximate", "pseudoinverse"), col=COL, lty=1, pch=21, bty="n")
par(op)


}
\references{
\url{http://stats.stackexchange.com/a/115477/10675}
}

