% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{plotStream}
\alias{plotStream}
\title{Stream plot}
\usage{
plotStream(x, y, order.method = "as.is", frac.rand = 0.1, spar = 0.2,
  center = TRUE, ylab = "", xlab = "", border = NULL, lwd = 1,
  col = rainbow(length(y[1, ])), ylim = NULL, ...)
}
\arguments{
\item{x}{A vector of values}

\item{y}{A matrix of data series (columns) corresponding to x}

\item{order.method}{Method of ordering y plotting order. One of the
following: \code{c("as.is", "max", "first")}. \code{"as.is"} - plot in
order of y column. \code{"max"} - plot in order of when each y series
reaches maximum value. \code{"first"} - plot in order of when each y series
first value > 0.}

\item{frac.rand}{Fraction of the overall data "stream" range used to
define the range of random wiggle (uniform distrubution)
to be added to the baseline 'g0'}

\item{spar}{Setting for smooth.spline function to make a smoothed version of baseline "g0"}

\item{center}{Logical. If TRUE, the stacked polygons will be centered
so that the middle, i.e. baseline ("g0"), of the stream is
approximately equal to zero. Centering is done before the
addition of random wiggle to the baseline.}

\item{border}{Border colors for polygons corresponding to y columns (will recycle) (see ?polygon for details)}

\item{lwd}{Border line width for polygons corresponding to y columns (will recycle)}

\item{col}{Fill colors for polygons corresponding to y columns (will recycle).}

\item{...}{Other plot arguments}
}
\description{
\code{plotStream} makes a "stream plot" where each \code{y}
series is plotted as stacked filled polygons on
alternating sides of a baseline.
makes a stacked plot where each
}
\examples{
x+y
}

