% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqRan.R
\name{seqRan}
\alias{seqRan}
\title{Create a sequence of defined length over range of a vector}
\usage{
seqRan(x, length.out = NULL, rel.ext = c(0, 0), na.rm = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{length.out}{Length of sequence to create, Defaults to length
of \code{x}.}

\item{rel.ext}{Relative extention of sequence limits. 
Default \code{rel.xt = c(0,0)} maintains original range of \code{x}}

\item{na.rm}{Logical. Should NAs be removed (see \code{\link[base]{range}})}

\item{...}{Additional parameters to pass to \code{\link[base]{seq}}}
}
\value{
Vector
}
\description{
Create a sequence of defined length over range of a vector
}
\examples{
x <- c(0,1,3,4)
seqRan(x) # default settings. Returns vector of length of x
seqRan(x, 10) # using original range of x
seqRan(x, 10, c(-0.1, 0.1)) # using extended range of x

}

